+++ docsig/messages.py
@@ -25,7 +25,7 @@ docsig.messages
 from ._message import Message as _Message
 from ._message import Messages as _Messages
 
-_TEMPLATE = "{code}: {description}"
+TEMPLATE = "{code}: {description}"
 _HINT_TEMPLATE = "hint: {hint}"
 
 
@@ -123,28 +123,28 @@ E = _Messages(
 
 # Exxx: Error
 # E1xx: Docstring
-E101 = str(E[101].fstring(_TEMPLATE))
-E102 = str(E[102].fstring(_TEMPLATE))
-E103 = str(E[103].fstring(_TEMPLATE))
-E104 = str(E[104].fstring(_TEMPLATE))
-E105 = str(E[105].fstring(_TEMPLATE))
-E106 = str(E[106].fstring(_TEMPLATE))
-E107 = str(E[107].fstring(_TEMPLATE))
-E108 = str(E[108].fstring(_TEMPLATE))
-E109 = str(E[109].fstring(_TEMPLATE))
-E110 = str(E[110].fstring(_TEMPLATE))
-E111 = str(E[111].fstring(_TEMPLATE))
-E112 = str(E[112].fstring(_TEMPLATE))
-E113 = str(E[113].fstring(_TEMPLATE))
-E114 = str(E[114].fstring(_TEMPLATE))
-E115 = str(E[115].fstring(_TEMPLATE))
-E116 = str(E[116].fstring(_TEMPLATE))
+E101 = str(E[101].fstring(TEMPLATE))
+E102 = str(E[102].fstring(TEMPLATE))
+E103 = str(E[103].fstring(TEMPLATE))
+E104 = str(E[104].fstring(TEMPLATE))
+E105 = str(E[105].fstring(TEMPLATE))
+E106 = str(E[106].fstring(TEMPLATE))
+E107 = str(E[107].fstring(TEMPLATE))
+E108 = str(E[108].fstring(TEMPLATE))
+E109 = str(E[109].fstring(TEMPLATE))
+E110 = str(E[110].fstring(TEMPLATE))
+E111 = str(E[111].fstring(TEMPLATE))
+E112 = str(E[112].fstring(TEMPLATE))
+E113 = str(E[113].fstring(TEMPLATE))
+E114 = str(E[114].fstring(TEMPLATE))
+E115 = str(E[115].fstring(TEMPLATE))
+E116 = str(E[116].fstring(TEMPLATE))
 
 # E2xx: Config
-E201 = str(E[201].fstring(_TEMPLATE))
-E202 = str(E[202].fstring(_TEMPLATE))
-E203 = str(E[203].fstring(_TEMPLATE))
-E204 = str(E[204].fstring(_TEMPLATE))
+E201 = str(E[201].fstring(TEMPLATE))
+E202 = str(E[202].fstring(TEMPLATE))
+E203 = str(E[203].fstring(TEMPLATE))
+E204 = str(E[204].fstring(TEMPLATE))
 
 # Hxxx: Hint
 # H1xx: Docstring